Add disallow license package options.
Provide an option to fail image download if a package, or license
exists in the given image.

The list packages option can now also be filtered by licenses and
package name.

The disallow package name can use the wildcard (*) to match a
name expression. For example *-mini can be filtered to prevent
mini packages in image.