Don't fail on missing metadata
The method get_image_packages_metadata looks up either a .report
or a .packages file from the given download URI. The information
if present is used in mash to check on package conditions and for
the build time of the image. In case no metadata information is
present the returned package dict is empty and the object instance
stored build time stays at its initial 'unknown' value. Not
raising an exception but sending a log warning has two effects
on mash:

1. An unknown buildtime causes a download of the image in any case
2. An empty packages information causes the optionally given
   package based conditions in the mash job description to be
   not verifiable

On the positive side any URL just providing an image can then
be consumed by mash with the mentioned restrictions. At the
moment the scope of mash is limited to OBS build images and
projects which has the "withreports" flags configured. Any other
source location is not consumable by mash and this commit
would allow to increase the scope for mash.